block deletion when team is in match

+++ volleyscope/server/api/team/deleteTeam.ts
@@ -1,6 +1,7 @@
 import { db } from "~/db/index";
 import { eq } from "drizzle-orm";
 import { teams } from "~/db/schema/players";
+import { matches } from "~/db/schema/match";
 
 export default defineEventHandler(async (event) => {
   const query = getQuery(event);
@@ -10,7 +11,13 @@ export default defineEventHandler(async (event) => {
     return { message: "Invalid Team ID" };
   }
   try {
-    await db.delete(teams).where(eq(teams.teamId, id));
+    const playerInfo = await db
+      .select()
+      .from(matches)
+      .where(eq(matches.teamId, id));
+    if (playerInfo === null) {
+      await db.delete(teams).where(eq(teams.teamId, id));
+    } else return { message: "used in match" };
   } catch (error) {
     console.log(error);
     return { message: "error" };
